requirements_simulator.txt
# Requirements for Pico W Simulator
# ติดตั้งด้วย: pip install -r requirements_simulator.txt

flask==3.0.0
paho-mqtt==1.6.1
ntplib==0.4.0
requests==2.31.0